Update on-and-on.scpt
Minor revision to exhibit "On and On" on Discord, rather than facebook. Ensures Discord is open at same time as Spotify, and sets the volume to zero. (I call this a gallery attendant quality-of-life feature)

Ensure that your Discord user profile has the Spotify integration activated, with the option to set Spotify as your status.

Ultimately, I don't like the status message on Discord as much aesthetically as posting the status to Facebook (as was the original intention of the piece), because Facebook showed the track being played at a glance, and discord requires clicking on the user's status to see that they are, in fact, listening to "Like a G6." But I'm never on Facebook and always on Discord, so there you go.

Done with artist's permission. Thank you Cory!